requirements.txt
python-dotenv>=1.0.0
fastmcp>=2.0.0
httpx>=0.27.0
starlette>=0.37.0
uvicorn>=0.30.0
pyjwt>=2.8.0
python-multipart>=0.0.9
itsdangerous>=2.1.0
